Set Stack id to None when deleted
If a Stack object has been removed from the database, we should reset its
ID to None so that if the object is used further, it will not be assumed
that it is represented in the database.

+++ heat/tests/utils.py
@@ -45,7 +45,8 @@ def stack_delete_after(test_fn):
             test_fn(test_cls)
         finally:
             try:
-                test_cls.stack.delete()
+                if test_cls.stack.id is not None:
+                    test_cls.stack.delete()
             except AttributeError:
                 pass
     return wrapped_test
